backend/requirements.txt
# fastapi==0.104.1
# uvicorn==0.24.0
# sqlalchemy==2.0.23
# psycopg2-binary==2.9.9
# pydantic==2.5.0
# pydantic-settings==2.1.0
# pydantic[email]==2.5.0
# python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4
# python-dotenv==1.0.0
# groq==0.4.1
# cors==1.0.1
# requests==2.31.0

fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.1
python-dotenv==1.0.0
groq==0.4.1
httpx==0.27.0
requests==2.31.0
email-validator